Restructure the autoconfigure module
