+++ hawkbit-autoconfigure/src/main/java/org/eclipse/hawkbit/autoconfigure/DefaultPropertiesAutoConfiguration.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.hawkbit.autoconfigure.conf;
+package org.eclipse.hawkbit.autoconfigure;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
