+++ hawkbit-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,18 +1,18 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.eclipse.hawkbit.autoconfigure.conf.DefaultPropertiesAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.ui.UIAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.security.SecurityAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.url.PropertyHostnameResolverAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.web.WebMvcAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.web.DDiApiAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.web.MgmtApiAutoConfiguration,\
 org.eclipse.hawkbit.autoconfigure.cache.CacheAutoConfiguration,\
 org.eclipse.hawkbit.autoconfigure.cache.DownloadIdCacheAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.event.EventPublisherAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.ddi.DDiApiAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.dmf.amqp.DmfApiAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.mgmt.ui.MgmtUiAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.mgmt.MgmtApiAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.repository.event.EventPublisherAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.repository.ArtifactFilesystemAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.repository.JpaRepositoryAutoConfiguration,\
 org.eclipse.hawkbit.autoconfigure.scheduling.AsyncConfigurerAutoConfiguration,\
 org.eclipse.hawkbit.autoconfigure.scheduling.ExecutorAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.amqp.DmfApiAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.security.InMemoryUserManagementConfiguration,\
-org.eclipse.hawkbit.autoconfigure.repository.ArtifactFilesystemAutoConfiguration,\
-org.eclipse.hawkbit.autoconfigure.repository.JpaRepositoryAutoConfiguration
+org.eclipse.hawkbit.autoconfigure.security.SecurityAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.security.InMemoryUserManagementAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.web.WebMvcAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.DefaultPropertiesAutoConfiguration,\
+org.eclipse.hawkbit.autoconfigure.PropertyHostnameResolverAutoConfiguration
